Zero shot module renamed and tooltips fixed on zero-shot page
