+++ src/app/zero-shot-details/zero-shot-details.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ZeroShotDetailsComponent } from './component/zero-shot-details.component';
-import { BaseModule } from './../base/base.module';
+import { BaseModule } from '../base/base.module';
